requirements.txt
Flask==2.3.2
python-docx==1.1.2
wordcloud==1.9.3
nltk==3.8.1
spacy==3.7.2
matplotlib==3.8.4
PyMuPDF==1.23.14